(QENG-7422) Add sleep to Unix Host Reboot
Prior to this commit the reboot method didn't wait before checking to
see if the host is back up after sending a shutdown.
This commit adds a sleep in order to give the host a little time to shut
down and reduce the likelyhood of hitting the host while it is in the
middle of shutdown and thus failing for a mysterious/hard to reproduce
reason.